+++ scripts/components/profiles/tooltip.component.js
@@ -1,5 +1,5 @@
 import TooltipTemplate from 'ejs!templates/shared/tooltip.ejs';
-import 'styles/components/shared/infowindow.scss';
+import 'styles/components/shared/info-tooltip.scss';
 
 export default class {
   constructor(className) {
@@ -10,7 +10,7 @@ export default class {
     this.infowindow.style.left = x + 'px';
     this.infowindow.style.top = y + 'px';
 
-    const template = TooltipTemplate({title: data.title, values: data.values});
+    const template = TooltipTemplate({ title: data.title, values: data.values });
     this.infowindow.innerHTML = template;
     this.infowindow.classList.remove('is-hidden');
   }