+++ CHANGELOG.md
@@ -1,5 +1,17 @@
+- profiles: includes map choropleth legend
+- profiles: includes the top sources' source switch.
+- tool: map dimensions groups are now collapsible
+- tool: map dimensions selecting is simpler (singl radio buttons column)
+- tool: fixed tool crash when links API fails
+- profiles: connects choropleth maps to real data in actors profile pages
+- home: only allow video playing on full screen
+- home: show screenshot instead of video when video playback has not started yet
+- tool: added tooltips on the map to show info that is otherwise displayed on the bottom of the sankey
+- tool: fixed water scarcity color scale
 - tool: added link tooltips resize by information
 - tool: added link tooltips quant units
+- profiles: actor municipalities/biome/state switcher
+- profiles: added line chart locations on actors page
 - profiles: added hover effects on shortcut buttons
 - profiles: companies exporting chart (mock)
 - tool entry in menu now triggers a submenu with two entry points